+++ src/built-ins/export/RequestSyslogger.js
@@ -1,7 +1,6 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { WallClock } from '@this/clocks';
 import { IntfRequestLogger } from '@this/net-protocol';
 import { BaseService } from '@this/sys-framework';
 
@@ -18,24 +17,28 @@ export class RequestSyslogger extends BaseService {
   // @defaultConstructor
 
   /** @override */
-  now() {
-    return WallClock.now();
-  }
-
-  /** @override */
-  async requestStarted(networkInfo_unused, timingInfo_unused, request) {
+  async _impl_handleEvent_requestStarted(request) {
     request.logger?.request(request.infoForLog);
+
+    return true;
   }
 
   /** @override */
-  async requestEnded(networkInfo, timingInfo, request, response) {
-    // Note: This call isn't supposed to `throw`, even if there were errors
-    // thrown during handling.
-    const resInfo = await response.getInfoForLog(
-      networkInfo.nodeResponse, networkInfo.connectionSocket);
+  async _impl_handleEvent_requestEnded(request, response, networkInfo) {
+    try {
+      const { connectionSocket, nodeResponse } = networkInfo;
+      const responseInfo = await response.getInfoForLog(nodeResponse, connectionSocket);
 
-    request.logger?.response(resInfo);
-    request.logger?.timing(timingInfo);
+      request.logger?.response(responseInfo);
+    } catch (e) {
+      // Shouldn't happen, but if it does, it's better to log and move on than
+      // to let the system crash. Note, in particular, the call to
+      // `getInfoForLog()` (above) is never supposed to throw, even if the
+      // request or response caused some sort of error to be thrown.
+      this.logger?.errorWhileLoggingRequest(e);
+    }
+
+    return true;
   }
 
   /** @override */
@@ -52,7 +55,6 @@ export class RequestSyslogger extends BaseService {
   async _impl_stop(willReload_unused) {
     // No need to do anything.
   }
-
 
   //
   // Static members