+++ src/net-protocol/export/IntfRequestLogger.js
@@ -1,8 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { Duration, Moment } from '@this/data-values';
-import { IncomingRequest, OutgoingResponse, TypeNodeRequest, TypeNodeResponse }
+import { IncomingRequest, OutgoingResponse, TypeNodeResponse }
   from '@this/net-util';
 import { Methods } from '@this/typey';
 
@@ -17,34 +16,14 @@ import { Methods } from '@this/typey';
  */
 export class IntfRequestLogger {
   /**
-   * Gets this instance's idea of what the current time is.
-   *
-   * @returns {Moment} The current time.
-   */
-  now() {
-    return Methods.abstract();
-  }
-
-  /**
    * Indicates to this instance that a request has started.
    *
    * @abstract
-   * @param {object} networkInfo Information about the network environment.
-   * @param {object} networkInfo.connectionSocket The socket (or socket-like
-   *   object) used by the lowest level of the connection that the request is
-   *   running on.
-   * @param {TypeNodeRequest} networkInfo.nodeRequest Low-level request object.
-   * @param {TypeNodeResponse} networkInfo.nodeResponse Low-level response
-   *   object.
-   * @param {object} timingInfo Information about request timing.
-   * @param {Moment} timingInfo.start The moment the request started getting
-   *   handled (or at least a reasonably close moment to that). This can be
-   *   expected to be a value returned from a call to {@link #now} on this
-   *   instance.
    * @param {IncomingRequest} request The incoming request.
+   * @returns {boolean} Whether or not the event was handled.
    */
-  async requestStarted(networkInfo, timingInfo, request) {
-    Methods.abstract(networkInfo, timingInfo, request);
+  async _impl_handleEvent_requestStarted(request) {
+    throw Methods.abstract(request);
   }
 
   /**
@@ -52,19 +31,18 @@ export class IntfRequestLogger {
    * has been sent).
    *
    * @abstract
-   * @param {object} networkInfo Information about the network environment. See
-   *   {@link #requestStarted} for details.
-   * @param {object} timingInfo Information about request timing.
-   * @param {Moment} timingInfo.start Same as with {@link #requestStarted}.
-   * @param {Moment} timingInfo.end The moment the request was considered
-   *   complete. This can be expected to be a value returned from a call to
-   *   {@link #now} on this instance.
-   * @param {Duration} timingInfo.duration The difference `end - start`.
    * @param {IncomingRequest} request The incoming request.
    * @param {OutgoingResponse} response The response that was sent (or at least
    *   attempted).
+   * @param {object} networkInfo Information about the network environment.
+   * @param {object} networkInfo.connectionSocket The socket (or socket-like
+   *   object) used by the lowest level of the connection that the request is
+   *   running on.
+   * @param {TypeNodeResponse} networkInfo.nodeResponse Low-level response
+   *   object.
+   * @returns {boolean} Whether or not the event was handled.
    */
-  async requestEnded(networkInfo, timingInfo, request, response) {
-    Methods.abstract(networkInfo, timingInfo, request, response);
+  async _impl_handleEvent_requestEnded(request, response, networkInfo) {
+    throw Methods.abstract(request, response, networkInfo);
   }
 }